refactor(gamelift-fleet): convert to libnuke resource format
